[UKET-181] 어드민 행사 등록 상태값 별 처리 (#81)
* [UKET-181] 검수완료 상태 변경 처리 추가

* [UKET-181] 등록완료 상태 변경 처리 추가

* [UKET-181] 행사완료 상태 변경 처리 추가

* [UKET-181] 컴파일 에러 수

* [UKET-181] 삭제 로직 추가

* [UKET-181] 인덱스 추가

* [UKET-181] 도메인 테스트 제거

* [UKET-181] 테스트 컴파일 제거

* [UKET-181] 테스트 수

* [UKET-181] lint

+++ src/main/kotlin/uket/api/user/response/TicketingResponse.kt
@@ -4,7 +4,7 @@ import uket.common.enums.BankCode
 
 data class TicketingResponse(
     val ticketIds: List<Long>,
-    val totalPrice: Int,
+    val totalPrice: Long,
     val depositUrl: String,
     val bankCode: BankCode,
 )
